completed backbone MVC refactor all MVC are partitioned in client and compatible with browserify
